+++ packages/design-tokens/src/build.js
@@ -1,15 +1,14 @@
-import { registerTransforms, transforms } from '@tokens-studio/sd-transforms';
-import StyleDictionary from 'style-dictionary';
 import { readFileSync } from 'fs';
+import { registerTransforms } from '@tokens-studio/sd-transforms';
+import StyleDictionary from 'style-dictionary';
+import { buildPath, prefix } from './config.js';
+import { calc } from './transforms/index.js';
+import { SDSTransformGroupDefault, SDSTransformGroupSASS } from './transformGroups.js';
 
-const { version } = JSON.parse(readFileSync('./package.json'));
-const buildPath = 'build/';
+const { author, name, version } = JSON.parse(readFileSync('./package.json'));
 
-// Set sds as our global variable prefix
-const tokenPrefix = 'sds';
-
-// Set up our own custom transform group
-const sdsTransformGroup = `${tokenPrefix}/tokens-studio`;
+// Add support for sds custom transforms
+StyleDictionary.registerTransform(calc);
 
 // Use @tokens-studio/sd-transforms from Tokens.Studio Figma plugin
 // Register them on StyleDictionary object
@@ -17,17 +16,14 @@ const sdsTransformGroup = `${tokenPrefix}/tokens-studio`;
 // @see https://github.com/tokens-studio/sd-transforms
 registerTransforms(StyleDictionary);
 
-// Register custom tokens-studio transform group
-// to add 'name/cti/kebab' for the token names
-StyleDictionary.registerTransformGroup({
-  name: sdsTransformGroup,
-  transforms: [...transforms, 'name/cti/kebab'],
-});
+StyleDictionary.registerTransformGroup(SDSTransformGroupDefault);
+StyleDictionary.registerTransformGroup(SDSTransformGroupSASS);
 
 // Set up our custom file header for better comments
 StyleDictionary.registerFileHeader({
   fileHeader: (defaultMsg) => [
-    `SICK Design System ${version}`,
+    `${name} version ${version}`,
+    `${author.name}`,
     ...defaultMsg,
   ],
   name: 'sdsHeader',
@@ -42,10 +38,11 @@ const sdsStyleDictionaryConfig = ({
         format: 'css/variables',
         options: {
           fileHeader: 'sdsHeader',
+          outputReferences: true,
         },
       }],
-      prefix: tokenPrefix,
-      transformGroup: sdsTransformGroup,
+      prefix,
+      transformGroup: SDSTransformGroupDefault.name,
     },
     js: {
       buildPath: `${buildPath}js/`,
@@ -65,7 +62,7 @@ const sdsStyleDictionaryConfig = ({
           },
         },
       ],
-      prefix: tokenPrefix,
+      prefix,
       transformGroup: 'tokens-studio',
     },
     scss: {
@@ -75,13 +72,14 @@ const sdsStyleDictionaryConfig = ({
         format: 'scss/variables',
         options: {
           fileHeader: 'sdsHeader',
+          outputReferences: true,
         },
       }],
       options: {
         themeable: true,
       },
-      prefix: tokenPrefix,
-      transformGroup: sdsTransformGroup,
+      prefix,
+      transformGroup: SDSTransformGroupSASS.name,
     },
   },
   source: ['tokens/**/*.json'],