+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Please have a look at the corresponding [packages](./packages) for a list of recent changes:
 
+- [@sick-design-system/css](./packages/css/CHANGELOG.md)
 - [@sick-design-system/design-tokens](./packages/design-tokens/CHANGELOG.md)
 - [@sick-design-system/eslint-config-sds](./packages/eslint-config-sds/CHANGELOG.md)
 - [@sick-design-system/icons](./packages/icons/CHANGELOG.md)